backend/requirements.txt
Flask==1.0.2
pymongo==3.8.0
python_bcrypt==0.3.2
